Clean up textStore.
